Split workspacetools.py, move some keymaps to keymaps.py
