+++ workspacetools/add_line2d.py
@@ -1,0 +1,26 @@
+from bpy.types import WorkSpaceTool
+
+from ..declarations import GizmoGroups, Operators, WorkSpaceTools
+from ..keymaps import disable_gizmos, tool_access, tool_select
+from ..stateful_operator.tool import GenericStateTool
+from ..stateful_operator.utilities.keymap import operator_access
+
+
+class VIEW3D_T_slvs_add_line2d(GenericStateTool, WorkSpaceTool):
+    bl_space_type = "VIEW_3D"
+    bl_context_mode = "OBJECT"
+    bl_idname = WorkSpaceTools.AddLine2D
+    bl_label = "Add 2D Line"
+    bl_operator = Operators.AddLine2D
+    bl_icon = "ops.gpencil.primitive_line"
+    bl_widget = GizmoGroups.Preselection
+    bl_keymap = (
+        *disable_gizmos,
+        *tool_select,
+        *tool_access,
+        *operator_access(Operators.AddLine2D),
+    )
+
+    def draw_settings(context, layout, tool):
+        props = tool.operator_properties(Operators.AddLine2D)
+        layout.prop(props, "continuous_draw")
