+++ workspacetools/add_line3d.py
@@ -1,0 +1,26 @@
+from bpy.types import WorkSpaceTool
+
+from ..declarations import GizmoGroups, Operators, WorkSpaceTools
+from ..keymaps import disable_gizmos, tool_access, tool_select
+from ..stateful_operator.tool import GenericStateTool
+from ..stateful_operator.utilities.keymap import operator_access
+
+
+class VIEW3D_T_slvs_add_line3d(GenericStateTool, WorkSpaceTool):
+    bl_space_type = "VIEW_3D"
+    bl_context_mode = "OBJECT"
+    bl_idname = WorkSpaceTools.AddLine3D
+    bl_label = "Add 3D Line"
+    bl_operator = Operators.AddLine3D
+    bl_icon = "ops.gpencil.primitive_line"
+    bl_widget = GizmoGroups.Preselection
+    bl_keymap = (
+        *disable_gizmos,
+        *tool_select,
+        *tool_access,
+        *operator_access(Operators.AddLine3D),
+    )
+
+    def draw_settings(context, layout, tool):
+        props = tool.operator_properties(Operators.AddLine3D)
+        layout.prop(props, "continuous_draw")
