+++ workspacetools/add_rectangle.py
@@ -1,0 +1,22 @@
+from bpy.types import WorkSpaceTool
+
+from ..declarations import GizmoGroups, Operators, WorkSpaceTools
+from ..keymaps import disable_gizmos, tool_access, tool_select
+from ..stateful_operator.tool import GenericStateTool
+from ..stateful_operator.utilities.keymap import operator_access
+
+
+class VIEW3D_T_slvs_add_rectangle(GenericStateTool, WorkSpaceTool):
+    bl_space_type = "VIEW_3D"
+    bl_context_mode = "OBJECT"
+    bl_idname = WorkSpaceTools.AddRectangle
+    bl_label = "Add Rectangle"
+    bl_operator = Operators.AddRectangle
+    bl_icon = "ops.gpencil.primitive_box"
+    bl_widget = GizmoGroups.Preselection
+    bl_keymap = (
+        *disable_gizmos,
+        *tool_select,
+        *tool_access,
+        *operator_access(Operators.AddRectangle),
+    )
